Cleanups as way of a review
Most significant change to the original set of work is to revert API interface
breakage in the client.  Instead of passing the transfer engine in explicitly,
make an internal reference.

For now, don't expose the TransferEngine object outside the helper functions
(`DoStat`, `DoGet`, `DoPut`); this may be revisited at some point in the future
but I don't want to expose the implementation details yet.