Unset http_proxy when it is set for tests
When http_proxy is set in a test - then a test using a cache is subsequently
run - the fake proxy name will cause a HTTP 500 to be returned by the cache
(as libcurl fails to resolve the proxy name).

This explicitly manages the http_proxy variable in a few other places where
we were accidentally depending on the ordering of tasks.