Directory listing changes in Pelican client
This is a different approach to the change of directory listings. Now,
for recursive downloads, we will query the director for a PROPFIND to
get the location of our "dirlisthost" (or origin). Then procede with the
normal recursive download code. If we receive a 405 from the director
doing this PROPFIND, we will revert back to the old way of doing things