Defer lookup of federation metadata until it is needed
The federation metadata was looked up at client configuration time,
even if the client is only going to do the equivalent of `--help`.

Defer the lookup until it's actually used.  This requires making sure
it's done once per config (using a `sync.Once`) and having a context
passed through at the call site (to ensure an appropriate timeout
occurs if needed).

Because the federation information is no longer fixed after the viper
config is loaded, we must access the corresponding variables through
an accessor function.  To avoid accidental re-introduction of direct
viper access, the variables are set to `direct_access: false`.

There's significant churn in the unit tests as we now must appropriately
initialize the client to reset the cache.  Where reasonable, contexts
are now passed through the call stack.