Plugin always return important classads
This commit adds a new "failTransfer" function that can be called from
the plugin on failure to populate the results channel with required
classads. This handles the case where we fail after setting up a
NewTransferJob or when doing a submit on the transferClient and before
we get any results back.