Recursive download plugin errors returned properly
There was a bug with the plugin when doing recursive downloads. If some
error happened such as a 404 on the directory we are requesting or a 500
which is doing the ?recursive query on a file instead of a directory,
PROPFIND would fail and the error was not making it back to the plugin.
Now, when we discover this error (eventually makes its way to RunMux()
in client/handle_http.go) we populate the results channel with the
failed result before closing it. This way, the plugin can see the
results and record the failure.

In addition, I added some unit tests toward these failures as well as
for recusive downloads for plugin in general.