Implement client direct reads
This implements direct read functionality within the client. Adds a
query parameter to add to the transfer url ?directread. When this is
set, the client will avoid using a cache and go directly to the origin
for its downloads. This PR also adds some unit tests for this
functionality with a few bug fixes along the way. This should work with
both the client and the plugin as well as with the ?pack and ?recursive
query.