NEAREST_CACHE preprend cache list when specified
To specify the local cache, users need to set the NEAREST_CACHE env
variable. In the old functionality, setting this means we do not look
for any other caches and just try to use this "preferred cache". Now, if
the specified NEAREST_CACHE has a '+' character at the end of it, it
will now append the "normal" list of caches behind the preferred cache
in the list of caches to try.