Shim in ability for the director to use load-aware cache redirects
Currently, the director blindly redirects to clients to the cache nearest to them
without any consideration of whether that cache may be overloaded. While we work
to make cache load information available in the Director, this adds a "load-aware"
load balancing function -- that currently assumes the same load for every cache.

When we have load information for each cache, this new feature should be able to
grab that information from the cache's serverAd and do something intelligent with it.
For now, I'm assuming load will be from [0,1] where 0 is no load and 1 is the highest.

This commit also makes it a bit easier to develop more load-balancing algorithms by
hooking up algorithm selection to a configurable parameter. For testing, we can
also randomly sort things.