Use HTTPS when the director is sending clients to Pelican servers
Previously, we determined the scheme based on whether the requested object
required an authorization token, but all Pelican servers should be HTTPS
all of the time. The only time we need to redirect to HTTP is when when the
client is fetching a public object from a legacy cache/origin.