Remove the need for =true for ?recursive
This removes the need for a value to the ?recursive query parameter, we
just check that it exists. Added a deprecation warning if a user
utilizes a value and informs the user the value will be ignored. In
addition, added the ability to use the ?recursive parameter outside of
the plugin to avoid any confusion.