Fix bug when director returns 404 for PROPFIND
Found a bug when we try to PROPFIND against an outdated director, the
director returns a 404 not found (because it redirects to a cache and
that cache cannot find the object to PROPFIND so it returns a 404).
When this happens the client returns just the error and not the
response. Therefore, the fix is to check if we fail querying the
director with a PROPFIND, we check if the namespace has a dirlisthost.
If dirlisthost is present, we continue and otherwise, we fail. We'll
need to eventually create a way to test against outdated directors but
for now, we'll stick to this band-aid.