Implement pelican object ls client command
New command added `pelican object ls` that lists the objects of the
specified dir in a namespace (or on the namespace itself). This commit
also comes with tests for this new command as well as tests for a helper
function to obtain 'dirlisthost'. The new command with a few sets of
flags that I felt were the most important from the standard 'ls' command