Throw clear error when InitClient() not called
The transfer engine in the client now throws a clear error if
config.InitClient() has not been called. Before this fix, if
InitClient() was not called we would get an error from the
TransferEngine about not having worker count set or other
unexpected/undesired behavior. Now, InitClient() checks if the
workercount has been set (sets a default value if not) and
NewTransferEngine() now throws an error if the client has not been
initialized.