Client de-duplicate cache entries from director
The Pelican client now will look out for duplicate caches to try
returned from the director. This is done with a new helper function that
will get the caches to try and check for any duplicates. If we find a
duplicate, it will not be added to the list of the caches to try. In
addition, there are various cleanups of old debug logging as well as
some new logging that seems cleaner without removing anything important.