Changed output for Mode and added json flag
There is now an ability to print out ls in json format. In addition,
worked around some output differently. The file mode will be printed in
string format in json output as well as long (I like the string format
but this is flexible if another format is more useful). Also messed with
the tests a little in the hopes it passed macOS failures.