Implementation of pelican object stat
The initial implentation of object stat command. This allows users to
run the command `pelican object stat` on an object in a namespace and
get the following information: Name, Size, ModTime, and IsDir. The only
current issue is that if we stat a directory we are currently unable to
give a size as the webdav client does not support it.