Increase `WaitUntilWorking` transfer timeout and adjust error message
When `WaitUtilWorking` was pointed at the OSDF Director (located in Madison)
from a distant server, it couldn't get a response in the previously-configured
50ms. After peeking at this function, it doesn't seem like there's a good reason
not to have a more permissive timeout.

I also adjusted the error message that was output from the previous circumstance.
Setting up a distant cache resulted in saying "Server director <url> did not startup
after 10s of waiting," which makes it seem like the cache is trying to start up a
director.