Fix plugin bug for incorrect start time
The plugin was reporting the start time of the transfer in the wrong
place. Originally, it was recorded right when we get the results for the
transfer (which is wrong). Now, we record the start time in the
TransferResults object so each transfer will have its own start time and
it will be properly recording in the plugin output ad.