Fix error msg, timeout for metadata
Made changes to the error message printed for metadata timeouts. For the
plugin the message was very verbose and contained some information that
is not very useful (especially towards users). This reformats the
message to contain just information that there was a timeout and how
long it took to timeout (10s). Also tweaked a function in plugin.go to
get rid of a non-used variable.

I did not think this needs testing since it is just changes to an error
message that is changing relatively often