Change various timeout settings to be durations, not integers
The various client timeouts were provided as integers (as the `duration`
configuration type didn't exist when they were added).  This migrates
everything to a duration instead.

Provides some level of backward compatibility; since viper interprets
a number without a suffix as a time in nanoseconds, we assume any duration
that's less than a microsecond was intended to be in seconds, not
nanoseconds.