Bug fix for '+' in NEAREST_CACHE
There was a bug when having a '+' character in the NEAREST_CACHE
list. The cache list would be populated correctly with the local cache
as the first item and the 'normal' cache list appended after it.
However, when calling sortAttempts() we would try to do a HEAD request
against the local cache (to sort our fastest caches first) but that
would fail due to an unrecognized scheme, causing the local cache to go to the end of the list of
attempts to try. This fix now looks at the scheme and if it is a
`unix://` scheme we skip the HEAD request. This also adds more tests to
ensure this does not happen.