Update client docs - new PR (#1381)
* Update pelican client documentation

More detailed and up to date client documentation, still a work in
progress and work needed for a "quick-start" page

* Updated documentation for client

Updated client-usage docs to be more up to date and include some
features. Also, added a quick-start guide for client usage.

* Updated client_usage docs based on feedback

* Update quick start guide based on feedback

* Updated documentation based on feedback

* Updated based on feedback

* Updates to client documentation

These changes include the renaming of the files to be more
descriptive/match the formatting of the other files. This also takes
into account the notes Andrew left and that we made during the
documentation planning meeting, converting the quickstart guide to a
getting started with client page.

* Address comments need discussion on namespaces

* Basic style formatting

* One pass for all code block spacing issues

* Address Andrew's comments

* Fix various small issues

* Address code review comments

---------

Co-authored-by: Joe Reuss <jrreuss@wisc.edu>