Client `stat` branch (#1400)
* Implementation of pelican object stat

The initial implentation of object stat command. This allows users to
run the command `pelican object stat` on an object in a namespace and
get the following information: Name, Size, ModTime, and IsDir. The only
current issue is that if we stat a directory we are currently unable to
give a size as the webdav client does not support it.

* Fix failing tests and add unit tests

Also, fixed linter issue for gosimple being unhappy with 'statResults'
being defined right before the channel 'resultsChan' (all within
statHttp function). Making the linter happy would mean adding more lines
of code than if we just keep it so I just moved it in the hopes the
linter can't find this issue again

* Some PR cleanup

Added comments and hopefully fixed gosimple linter error

* Fix old code review comments

* Remove unused function arg

* Remove shadowed name for token package

* Merge branch 'main' of github.com:PelicanPlatform/pelican into client-stat-hm

* Address code review comments from Justin

* Remove path in the collection URL

* Return fileInfo instead of printing the result in the client library

* Remove unused code and fix test failure

* Fix uncaught merge conflict

---------

Co-authored-by: Joe Reuss <jrreuss@wisc.edu>