Gut client of Topology integration
The purpose of this was to remove Topology integrations from the client and
rely instead solely on the director. This was accomplished by getting rid
of the `namespaces` package.

The big picture change here is that instead of having the client construct
and pass around namespace objects, it now constructs and passes around a
director response object, which is designed to have all the information needed
to perform various client actions.

While working on this, I realized the Director also stopped responding with the
`base-paths` key in the `X-Pelican-Token-Generation` header, so I lumped that
into this (which I kind of had to do for some of the changes to make sense).
That issue is tracked [here](https://github.com/PelicanPlatform/pelican/issues/1540),
but I'll note this is only a partial, temporary fix to stop potential bleeding --
more thought should be given to it longterm.

While reviewing, it's safe to ignore the many thousands of lines of deleted
code from the `namespaces` module. This is all legacy, and the goal of this
was to remove it all and replace it with director equivalents.

I also fixed up a few duplicated Director requests I found, such as was the case
with recursive downloads that called the Director twice -- once to construct
namespace info, and once to get the collections URL. Now this is all done in the
same query.