Don't start director on invalid sort method and fix cyclic import issue
This commit started out as "Don't start the director if there's an invalid
sort method in the configuration" because I observed that a vad value would
let the Director run, but would respond to all client requests with an error.
Then I hit a cyclic import issue and went on a tangent that was light in
content (mostly changing `config.ServerType` to `server_structs.ServerType`),
but heavy in files changed.

To get around the issue, I moved the `ServerType` struct and its methods to
`server_structs/generic_server.go`. I left `enabledServers` et al in
config.go, because these pertain specifically to which servers are enabled
during the config initialization process.