Consolidate sort's `candidateLimit` and Director's `serverResLimit`
As implemented, the adaptive sort protocol would only return `candidateLimit = 3`
servers for the Director to include in its redirect response. However, we already
have `serverResLimit = 6`, which configures the max number of servers with which
to supply a client request. Since these two variables have the same effect (i.e.
limiting the number of sorted servers sent to clients by the Director), why not
make them the same variable?

That's what this commit does :)