Selectively enable client X509 authentication
When a requested object matches a pre-configured set of paths, then
indicate to the cache it should turn on client X509 authentication.

This fixes support for legacy OSDF namespaces which may require the
cache to use client X509 authentication when requesting from the
origin because a token is not otherwise available.