Cache the object existence responses from the service
This PR allows the object existence response to be cached at the
director, preventing the remote service from getting repeated queries
for the same object.

It also splits the parameters enabling the director cache so we can
control separately the origin and cache behavior.

Finally, if there's a single origin, the director can assume it can
elide the existence check: assume that an object always exists at
the one origin that can provide it.