Add timed-wait functionality to the errgroup
This is a fork of the upstream errgroup code with an additional method
that allows a timed wait.

Those using the `TryGoUntil` method will block waiting for the ability
to execute in the errgroup _or_ the provided context is canceled.  This
allows callers to have a third option between "never block" and "block
indefinitely".