Fix topology downtime integration by using downtime URL and not ns JSON
This changes the overall approach for grabbing downtime information from Topology.
Previously, we generated the list of downed servers by getting a diff of the servers
returned in the namespaces JSON when querying that endpoint with `?include_downed=<0/1>`.
The issue with this approach is that Topology services marked as Pelican services are
explicitly excluded from the returned JSON.

However, Pelican services that are recorded in Topology are still included in the downtime
URL, so this new approach grabs the list of resource names from that instead. That was
probably the right approach all along, but topology endpoints remain inscrutable ¯\_(ツ)_/¯

This also corrects a documentation issue that claimed filtered servers could be added
in the Director using their hostname. In fact, they need to be filtered using their
site name!