Convert the token from a string to an object
The token tracked internally through a transfer is now a standalone struct
instead of a string.  The `get` method that produces the string needed for
transfers will check the validity of the current token and, if it's expired,
iterate through the environment to find a new one.

The goal is long-lived transfers will be able to acquire a fresh token, as needed,
from their environment (or from the token issuer) instead of using the expired one.