Fix race condition on recursive transfers
If a recursive transfer has more than one resulting object movement-- and
the transfers all finish before the directory walk has completed -- then the
overall transfer job was never be marked as completed.

This adds a check for job completion after the end of the directory walk
completion.  It also refactors the active transfer cleanup logic into a
standalone function to avoid a large copy/paste of code.