Plumb HTTPS backend through CLI and adjust it to use full storage prefixes
This PR accomplishes two things -- first, it plumbs a few of the https backend
components through the CLI, allowing users to serve minimal https origins without
writing a configuration yaml.

Second, it adjust the https backend to use storage prefixes. Consider the following
configuration:
```
Origin:
  StorageType: "https"
  HttpServiceUrl: "https://data.lhncbc.nlm.nih.gov/public"
  Exports:
    - StoragePrefix: "/Tuberculosis-Chest-X-ray-Datasets/Montgomery-County-CXR-Set/MontgomerySet/CXR_png"
      FederationPrefix: "/my-prefix"
```

The adjustments cause a request for object `/my-prefix/MCUCXR_0005_0.png` to be converted to
a libCurl request in the Origin for `https://data.lhncbc.nlm.nih.gov/public/Tuberculosis-Chest-X-ray-Datasets/Montgomery-County-CXR-Set/MontgomerySet/CXR_png/MCUCXR_0005_0.png`

While we don't yet support multiple exports for the https backend, I think this configuration
affords us maximum flexibility for the future where we do by allowing us to set the service URL
(every request to the origin uses this as the base URL), a storage prefix (each namespace can then
carve out some section of data hosted by the service url), and the typical federation prefix mapping
(i.e. strip the fed prefix from the object and tack that value to the end of the service url + storage prefix).

While this is a slight change in how an http origin would be configured, I'm not worried about backwards compat
because a) we never documented the limitations of this, even within our own codebase and b) this makes the
https backend conformant with the way we use storage prefixes in every other backend.