Assign bounded random coord for unknown client addresses
This assigns a random lat/long bounded within the contiguous US whenever the Director
is unable to determine where the client is coming from. We're moving to this because
the next step is to cache the assignment so that clients who fall into this trap can
improve follow-up reads through Pelican caches, and because most of our client base
falls within that box (and don't want to be redirected to caches in the UK, Russia, or
Singapore).