Null the token generator pointer for public reads
Prior to the new token generator object, an empty / missing token
was a no-op.  After it, the token generator could potentially try
to acquire a token.  Even though it was non-fatal when no token
was available, the error messages were awfully confusing.

Now, if no token is required, we explicitly nil out the token
generator object, preventing the request.