Fix nil pointer bug caused by trying to list namespaces with no coll URL
Previously, the code in `listHttp` didn't check whether or not the director response it had
actually contained a valid collections URL. Now it checks whether the URL is populated,
and if not returns an error with a hint.

I've also added very basic unit test coverage for this, but noticed both `listHttp` and
`statHttp` could use a lot more. Rather than do that in this bug fix PR, I'll open a followup
ticket that targets only these two functions for unit test development.