Unmarshal Capabilities correctly in all structs that use them
Instead of targeting NamespaceAdV2 like I did previously, this change moves
the unmarshal logic to the level of the Capabilities struct. Why I didn't do
that in the first place? I couldn't tell ya

In addition, I now test that JSON for ServerAd and OriginAdvertiseV2 (which may
also arrive to the Director with old/new capabilities) unmarshal correctly.