Perform standard federation metadata discovery for determining fed issuer
Between the writing and compiling out of this code, we've hidden direct access
through the param package to the 'Federation.DirectorUrl' key. Instead, this now
calls `config.GetFederation()`, which tries to perform metadata discovery in many
cases.

To fix this, the commit adjusts the way we grab this information, as well as fixes
tests that now mock the metadata discovery as well.