Create Lotman-enabled server binary and plumb through to prod cache containers
This creates a second `pelican-server` binary that passes the `lotman` build flag to
go, thereby enabling the previously-shimmed Lotman wrapper package. Since we only
envision this being used in caches (at least for the time being), this new binary
only replaces the binaries for the `cache` and `osdf-cache` targeted Docker builds.

Along with this, the entrypoints for these containers have been modified to accept
`pelican-server` and `osdf-server` arguments.