Attempt running dual test suites -- one for server binary, one for regular
Github Actions workflows are not my forte, but the attempt here is to re-use the
workflow file (now test-template.yml) to duplicate all the tests but with the new
binary and with the lotman tests.

Since GHA doesn't provide a way to mock the action environment in a local dry-run,
I'll be developing through commits/pushes. Let's see how it goes...