Move /pelican/monitoring into the "namespace" directory
The /pelican/monitoring namespace is where we write various test files
that are used to determine whether an XRootD service is currently healthy.
After updating our default locations for various cache directories, this
was missed, and the monitoring namespace was written to an inaccessible
location. This commit corrects the locations.