Add new config infrastructure for dropping to `pelican` user
This introduces new parameters for dropping to a lower-privileged
`pelican` user when running as root and the first round of directory
ownership changes.

The goal is that, when run as root and drop-priv is enabled, all
Pelican directories and config files are owned by the pelican user.
Otherwise, they'll continue to be owned as root.