Run release action against triggering tag instead of checking out branch
Previously, the action would trigger on a tag like `v7.12.3` and would then use
regex to checkout `v7.12.x`, which it would use for generating goreleaser artifacts.

If I'm parsing things correctly, this means that creating a tag and then adding extra
commits on top of 7.12.x _could_ cause us to build the wrong thing.

Instead, this just checks out whatever github ref triggered the action and builds that.