If all possible origins return a 404, declare the object not found
Slightly tweak the error handling logic so if one-or-more origin
returns a 404 for an object (and none return a success or an
indeterminate failure like 403 / 500) then declare the object as
not existing.

Previously, if all origins returned a 404, then the logic declared
the query as having an insufficient number of responses and redirected
the client to one of the origins that just provided the 404.