Use alma9 container for building prod pelican-server binaries
Prior to the introduction of the pelican-server binary, pelican was
always statically-linked. However, pelican-sever needs to be dynamic so
it can access the lotman shared library.

This caused an issue in our production cache containers because the
pelican-server binary was built in an alpine container (inherited from
the old goreleaser container) which uses libmusl for linking, but then
copied into an Alma container in the final stage, where libmusl is
unavailable.

This this fix, the base container used to build pelican binaries is switched
from the goreleaser container to a raw alma9 container. A side effect of this
is that we now have to install a few things we previously didn't. Another
thing I can't readily explain is that making this change appears to have
altered the path these binaries are built under from
`/pelican/dist/pelican_linux_amd64_v1/pelican`
to
`/pelican/dist/linux_amd64/pelican_linux_amd64_v1/pelican`

This doesn't appear to be an issue and was accounted for by changing a few
paths in the Dockerfile. Famous last words?