Memory usage stress test
This additional stress test generates a large number of object requests
against the director and measures the golang heap usage before/after
the requests.  The goal is to ensure the memory usage stays between
known fenceposts, hopefully acting as an early warning that the
stat cache code has gone wrong.