Fix race condition in `TestCache`
The `TestCache` test incremented a counter in one goroutine during
HTTP request processing and read it from the main test thread to
see if a request was finished.

Since this was not an atomic counter, the change in the counter could
get propagated at any arbitrary time in the future, meaning the latest
value need not be visible to the main thread.  This PR switches to
an atomic, fixing the race condition.

Additionally, set a TTL in the mock cache to a large value to avoid
potentially triggering a preemptive cache refresh.

This test has been flakey; hopefully this fixes the issue.