Adjust Client/Director retry logic to use configurable param
Now the client will try to detect two situations that may call for retries:
1) The "director's" response is missing the `Server: pelican/` header AND
there's a status code indicating some error. This means we're likely hitting
a proxy and should retry until the Server header reappears
2) The Server header is correct, but the Director responds with an HTTP 429
(too many requests) indicating it just rebooted and can't issue the requested
redirect, perhaps because it's waiting for a service to re-advertise

In both cases, the retry logic uses a configurable number of retries (where plugin
detection doubles the number of retries), and the retry frequency is backed off
with each unsuccessful attempt.