Accept 200 or 201 on uploads
The next version of XRootD will respond with a 201 (Created) on successful
upload (which is the correct status code).  Unfortunately, `pelican`
had a hard check for 200 (OK) instead.

Accept either.  Additionally, use the symbolic name for these so we
can better `grep` through the source code to find use of particular
status codes.