Restore the behavior of the keygen tool
In the branch for multi-key support, the keygen tool was changed
to create keys in the key directory instead of the working directory.

While that might have been a better original design, I don't want to
switch the semantic (and particularly the documentation!) at this point.

Restoring the original behavior.