Always set the audience
The audience setting in the generated scitokens.cfg only was firing
based on various configuration conditionals that might not always
be true (and, potentially, was set twice).

No matter what issuer is configured, the audience should be set.
Hoist the code to ensure it's always set -- and at most once.