Release v7.12.1
One note -- we know this new version introduces a directorTest bug
in the caches, but we're prioritizing other bugfixes while we work on
that. This release finally picks up xrdcl-pelican v1.0.2, which fixes
various segfaults on non-existent files, and also has fixes for cache
auth/scitokens files that update on startup, but fail to be otherwise
maintained.