Install pelican dependencies in final stage container
We've run into a lot of issues with the way we install dependencies in our
production container, including the inability to determine versions for
what's installed, and shared objects that are copied without bringing
along their dependencies.

Instead, we should just install whatever we can directly as an RPM into
the final-stage container. This solves both issues in one fell swoop
of a Pelican's beak.