Generate new v7.12.3-rc.0 to pick up bugfix for xrootd from osg-testing
We want to ship our containers with xrootd 5.7.2-1.5 from osg-testing, which
has a patch for a nil pointer bug affecting some caches. I was having issues
with yum metadata, so I also added a `--refresh` to the Dockerfile so I don't have
wait for the metadata to expire before generating the patch.