Add "federation token" endpoint to the Director
This new endpoint allows caches or origins to get a federation-signed
token to provide other federation services that may require proof of
federation membership.

The immediate usecase is for caches to provide Origins that haven't
enabled DirectReads, although I think this may also be useful for a
future where we support third-party copies between origins. I tried
to add a little bit of flexibility to the endpoint to support that use
case, even though we don't have the immediate need.

The authz scheme here is that a server requesting a fed token must provide
its hostname, server type and a valid advertisement token to the Director.
The Director then verifies the advertisement token, which is de facto
proof of federation membership, and if valid issues a personally-signed
token.