Enable PROPFIND of Director health test API
The cache's `xrdcl-pelican` plugin switched from doing a HEAD request against
requested resources to a PROPFIND request, the answer to which is used by the
cache to determine whether something is an object or a collection. That's
hunky dory, except that it broke the pseudo Director origin responsible for
generating health test files, because the Director's health test API wasn't
handling the propfinds.

This commit adds PROPFIND handling and generates the XML needed by the cache to
understand the response. Note that I constructed the XML by pinging a cache with
a real PROPFIND and then templating the response with file name, size, and time
stamp.

One item remaining here is an actual E2E test of the setup -- while the unit tests
are passing just fine, they passed just fine when everything broke... a better
setup is forthcoming.