Make a common base ad for all server types; add director ad type
The common base includes the name -- but also the version number,
a process/instance identifier, and a generation number.

The instance ID and generation number provide a total ordering for
all ads generated by the name/type of ad.  This will allow a process
to compare any two ads and determine which one is the "newer" of the
two.  This will allow the director to keep the right ad in memory
even if it receives them out-of-order.