Add multi-director detection and ad forwarding
Provide a framework for any service to detect the presence of multiple
director services in a federation, whether from static configuration,
from metadata discovery, or from querying known directors.  All the
sources are merged together to create one set of known directors.

Once the directors are known, and advertisements from the service
are sent to all directors directly from the service.

Additionally, the directors will forward the ads they receive from
a service to all other known directors.  This prevents a director
from missing updates just because the service hadn't discovered it
yet.

The directors will compare the generation and instance IDs of the
incoming ads to remove duplicates and ensure they only keep the
latest copy of a given service description.

A simple unit test is included.