Fix bug in adaptive sort's stochastic sort function
There was a bug previously that assumed all incoming weights were positive.
However, some of the weights may be negative, which is a trick used to make
sure the weights are always sorted at the end of a list. When negative weights
made it into adaptive sort, it triggered an infinite loop.

This fixes the bug by normalizing the weights in a way that guarantees negative
weights are treated as the smallest possible ranges.

Additionally, rather than generate random numbers over and over and over only
to see them consistently falling in an already-visited range, this changes the
function to remove and rescale ranges as we go. One consequence of this is that
it no longer makes sense for the stochastic sort function to return the weights
because these weights only make sense in the context of the exact set of ranges
that existed when the weight was created.